Prefer rule over tree indexing
If we start indexing into a tree; the intermediate value is represented
by a `TreeM` constructor.  We then try to avoid turning this `TreeM`
into a proper object, since doing so means fully evaluating it, which
isn't always necessary.

However, this optimization should not trigger if there is a rule at that
point in the tree; since then we should index into the rule and not the
tree.