Type error when assigning satured call
You can use functions that take arguments, such as `array.concat` in two ways:

1.  `array.concat([1], [2], out)`
2.  `out = array.concat([1], [2])` (alternatively using `:=`)

This makes you wonder what the _return type_ of `array.concat([1], [2], out)`
is.  We used to have a boolean here, that returned whether or not `out` unified
with the result.  However, that allows us to write code as:

    x = array.concat([1], [2], out)

Which is almost always an arity mistake on the programmer's side.  This patch
changes the return type of saturated calls to the _void_ type, so it can't be
assigned or used.