Fix issue in how and when we ground rule indices
When we have a rule with an index, e.g.:

    match[policy] {
        ...
    }

It's often a good idea to evaluate `policy` first.  That way we can bind that
free variable to whatever the rule is called as.  This gives us instant lookup
if we're calling it as `match["key"]`.

However, if there's a more complicated expression in the index:

    match[policy.name] {
        ...
    }

We must not evaluate the index until after we've evaluated the rule body.

Furthermore, we ensure that we're always returning a grounded `Value` from rule
definitions, rather than a possibly ungrounded `Mu`.

Fixes #233