Support using data imports as variables
If there's a data import, e.g.

    import data.foo

`fregot` only supported using `foo` as if it were a _package_ (e.g. `foo.bar`),
not as a _rule_ directly (i.e. `foo`).

There's really not much of a distinction between the two from a "data tree"
perspective, but it's helpful to keep the distinction anyway because it improves
error messages.  This patch adds support for the latter.

Fixes #238