Delay union on array types
The typechecker is a bit to eager in #292 to turn something like:

    array<A> | array<B>

Into:

    array<A|B>

In most cases this is fine, but sometimes you want the former, especially if
you're using an array as a tuple where the first field may be a discriminator.

This PR fixes that by not unifying the arrays, and just deconstructing it
later when necessary.